Moved the fireworks effect into a separate file to make it easier for adding more framework onboarding pages
