+++ apps/webapp/app/routes/_app.orgs.$organizationSlug.projects.$projectParam.setup.nextjs/route.tsx
@@ -1,7 +1,5 @@
 import { ChatBubbleLeftRightIcon, Squares2X2Icon } from "@heroicons/react/20/solid";
-import { useRevalidator } from "@remix-run/react";
-import { useEffect, useState } from "react";
-import { useEventSource } from "remix-utils";
+import { useState } from "react";
 import invariant from "tiny-invariant";
 import { Feedback } from "~/components/Feedback";
 import { PageGradient } from "~/components/PageGradient";
@@ -23,11 +21,12 @@ import { Paragraph } from "~/components/primitives/Paragraph";
 import { RadioGroup, RadioGroupItem } from "~/components/primitives/RadioButton";
 import { StepNumber } from "~/components/primitives/StepNumber";
 import { useAppOrigin } from "~/hooks/useAppOrigin";
+import { useProjectSetupComplete } from "~/hooks/useProjectSetupComplete";
 import { useDevEnvironment } from "~/hooks/useEnvironments";
 import { useOrganization } from "~/hooks/useOrganizations";
 import { useProject } from "~/hooks/useProject";
 import { Handle } from "~/utils/handle";
-import { projectSetupPath, projectStreamingPath, trimTrailingSlash } from "~/utils/pathBuilder";
+import { projectSetupPath, trimTrailingSlash } from "~/utils/pathBuilder";
 
 type SelectionChoices = "use-existing-project" | "create-new-next-app";
 
@@ -36,77 +35,14 @@ export const handle: Handle = {
 };
 
 export default function SetupNextjs() {
-  const project = useProject();
   const organization = useOrganization();
+  const project = useProject();
+  useProjectSetupComplete();
   const devEnvironment = useDevEnvironment();
   const appOrigin = useAppOrigin();
-
   const [selectedValue, setSelectedValue] = useState<SelectionChoices | null>(null);
 
   invariant(devEnvironment, "devEnvironment is required");
-
-  const revalidator = useRevalidator();
-  const events = useEventSource(projectStreamingPath(project.id), {
-    event: "message",
-  });
-
-  useEffect(() => {
-    if (events !== null) {
-      // This uses https://www.npmjs.com/package/canvas-confetti
-      if ("confetti" in window && typeof window.confetti !== "undefined") {
-        const duration = 3.5 * 1000;
-        const animationEnd = Date.now() + duration;
-        const defaults = {
-          startVelocity: 30,
-          spread: 360,
-          ticks: 60,
-          zIndex: 0,
-          colors: [
-            "#E7FF52",
-            "#41FF54",
-            "rgb(245 158 11)",
-            "rgb(22 163 74)",
-            "rgb(37 99 235)",
-            "rgb(67 56 202)",
-            "rgb(219 39 119)",
-            "rgb(225 29 72)",
-            "rgb(217 70 239)",
-          ],
-        };
-        function randomInRange(min: number, max: number): number {
-          return Math.random() * (max - min) + min;
-        }
-        // @ts-ignore
-        const interval = setInterval(function () {
-          const timeLeft = animationEnd - Date.now();
-
-          if (timeLeft <= 0) {
-            return clearInterval(interval);
-          }
-
-          const particleCount = 60 * (timeLeft / duration);
-          // since particles fall down, start a bit higher than random
-          // @ts-ignore
-          window.confetti(
-            Object.assign({}, defaults, {
-              particleCount,
-              origin: { x: randomInRange(0.1, 0.4), y: Math.random() - 0.2 },
-            })
-          );
-          // @ts-ignore
-          window.confetti(
-            Object.assign({}, defaults, {
-              particleCount,
-              origin: { x: randomInRange(0.6, 0.9), y: Math.random() - 0.2 },
-            })
-          );
-        }, 250);
-      }
-
-      revalidator.revalidate();
-    }
-    // WARNING Don't put the revalidator in the useEffect deps array or bad things will happen
-  }, [events]); // eslint-disable-line react-hooks/exhaustive-deps
 
   return (
     <PageGradient>
